+++ Validators/FatturaOrdinariaValidator.cs
@@ -106,7 +106,11 @@ namespace FatturaElettronica.Validators
             if (cessionario.IdFiscaleIVA.IsEmpty() && cessionario.CodiceFiscale == cedente.CodiceFiscale)
                 // Vedi https://github.com/FatturaElettronica/FatturaElettronica.NET/issues/388
                 return true;
-            
+
+            if (string.IsNullOrEmpty(cessionario.CodiceFiscale) &&
+                cedente.IdFiscaleIVA?.ToString() == cessionario.IdFiscaleIVA?.ToString())
+                return true;
+
             return cedente.IdFiscaleIVA?.ToString() == cessionario.IdFiscaleIVA?.ToString() &&
                    cedente.CodiceFiscale == cessionario.CodiceFiscale;
         }